uv-install-wheel: Split installation logic and link logic
uv-install-wheel had the logic for laying out the installation and for linking a directory in the same module. We split them up and tighten the crate's interface to only expose top level members.

No logic changes, only moving code around.
