+++ crates/uv-bench/benches/uv.rs
@@ -92,7 +92,7 @@ mod resolver {
     use uv_dispatch::{BuildDispatch, SharedState};
     use uv_distribution::DistributionDatabase;
     use uv_distribution_types::{DependencyMetadata, IndexLocations};
-    use uv_install_wheel::linker::LinkMode;
+    use uv_install_wheel::LinkMode;
     use uv_pep440::Version;
     use uv_pep508::{MarkerEnvironment, MarkerEnvironmentBuilder};
     use uv_platform_tags::{Arch, Os, Platform, Tags};
