+++ crates/uv-installer/src/installer.rs
@@ -9,7 +9,7 @@ use tracing::instrument;
 use uv_cache::Cache;
 use uv_configuration::RAYON_INITIALIZE;
 use uv_distribution_types::CachedDist;
-use uv_install_wheel::{linker::LinkMode, Layout};
+use uv_install_wheel::{Layout, LinkMode};
 use uv_python::PythonEnvironment;
 
 pub struct Installer<'a> {
@@ -34,7 +34,7 @@ impl<'a> Installer<'a> {
         }
     }
 
-    /// Set the [`LinkMode`][`uv_install_wheel::linker::LinkMode`] to use for this installer.
+    /// Set the [`LinkMode`][`uv_install_wheel::LinkMode`] to use for this installer.
     #[must_use]
     pub fn with_link_mode(self, link_mode: LinkMode) -> Self {
         Self { link_mode, ..self }
@@ -158,9 +158,9 @@ fn install(
 ) -> Result<Vec<CachedDist>> {
     // Initialize the threadpool with the user settings.
     LazyLock::force(&RAYON_INITIALIZE);
-    let locks = uv_install_wheel::linker::Locks::default();
+    let locks = uv_install_wheel::Locks::default();
     wheels.par_iter().try_for_each(|wheel| {
-        uv_install_wheel::linker::install_wheel(
+        uv_install_wheel::install_wheel(
             &layout,
             relocatable,
             wheel.path(),
